requirements.txt
numpy
nltk
flask
Bio
neurosynth
cognitiveatlas
gitpython
